Implement basic GRANT statement. #1860
Add grammar for simple grant statement. DATABASE level permissions only,
and privileges restricted to ALL (in grammar), or READ|WRITE (in
execution).

There's a decent list of TODOs in both sql/parser/sql.y and
sql/grant.go, some of which are open questions.

The parser part is the one that should be carefully examined.
For example, I know that the grantee_list should not be a list of 'name'
types. There are probably other issues.

Once this is in acceptable shape, I'll do the same for REVOKE, both of
which are needed before we start enforcing permissions.